backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pandas==2.2.3
python-dotenv==1.0.1
langchain-google-genai==2.0.7
pydantic==2.10.3
langchain-core==0.3.29
